Refactor reusable code into own component
